model/requirements.txt
Flask-SQLAlchemy
psycopg2-binary
scikit-learn
matplotlib
requests
numpy
pandas